successfully added conversation handler that takes the allows action on entered text and returns to main menu
